新功能tps

+++ src/main/kotlin/com/xiaoyv404/mirai/app/minecraftServer/Performance.kt
@@ -6,7 +6,7 @@ data class Performance(
     val timestamp: Long,
     val timestamp_f: String,
 //    val playersOnline:List<List<Number>>,
-    val tps:List<List<Long>>,
+    val tps:List<List<Float>>,
 //    val cpu:List<List<Number>>,
 //    val ram:List<List<Number>>,
 //    val entities:List<List<Number>>,
